requirements.txt
# python package requirements file

# Note to our Learners: Our courses are created at a specific point in time. The models and libraries used will naturally change over time.
# If you download and run the notebooks locally, you should create a Python virtual environment  
# with the Python version matching the one used in the course and use this requirements.txt file to install the required Python packages.
# Due to the wide diversity of platforms and environments, DeepLearning does not explicitly support running locally. However, if you encounter
# issues, the course support channel is a good place to ask questions.


# note which revision of python, for example 3.9.6
# in this file, insert all the pip install needs, include revision
# python 3.11

langchain==0.3.18
langchain-openai==0.3.5
langchain-community
langgraph==0.2.72
langmem==0.0.8
python-dotenv==1.0.1
IPython
sentence-transformers
